updated logic
